+++ src/main/java/com/codigo/spring/repository/AerolineaRepository.java
@@ -3,5 +3,5 @@ package com.codigo.spring.repository;
 import com.codigo.spring.entity.AerolineaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AerolineaRepository extends JpaRepository<AerolineaEntity, Integer> {
+public interface AerolineaRepository extends JpaRepository<AerolineaEntity, Long> {
 }
