+++ src/main/java/com/codigo/spring/repository/PasajeroRepository.java
@@ -3,5 +3,5 @@ package com.codigo.spring.repository;
 import com.codigo.spring.entity.PasajeroEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PasajeroRepository extends JpaRepository<PasajeroEntity, Integer> {
+public interface PasajeroRepository extends JpaRepository<PasajeroEntity, Long> {
 }
